Refactors ShoeboxDevModule Fixes AmazonInstanceInfo binding issues in SearchApplication, Moves MainSearcherTest to Search Module
