+++ kifi-backend/modules/common/app/com/keepit/common/cache/FortyTwoCache.scala
@@ -314,4 +314,3 @@ abstract class CacheModule(cachePluginModules: CachePluginModule*) extends Scala
     cachePluginModules.foreach(install)
   }
 }
-
