+++ kifi-backend/modules/common/app/com/keepit/dev/DevCommonModule.scala
@@ -12,12 +12,12 @@ import akka.actor.ActorSystem
 import play.api.Play
 import com.keepit.common.plugin.SchedulingProperties
 import com.keepit.common.analytics.Event
-import com.keepit.common.cache.DevCacheModule
+import com.keepit.module.DevDiscoveryModule
 
 class DevCommonModule extends ScalaModule with Logging {
   def configure() {
     install(new S3DevModule)
-    install(new DevCacheModule)
+    install(DevDiscoveryModule())
   }
 
   @Singleton
@@ -32,20 +32,4 @@ class DevCommonModule extends ScalaModule with Logging {
     }
   }
 
-  @Provides
-  @AppScoped
-  def actorPluginProvider: ActorPlugin =
-    new ActorPlugin(ActorSystem("shoebox-dev-actor-system", Play.current.configuration.underlying, Play.current.classloader))
-}
-
-class FakeEventPersisterImpl @Inject() (
-  system: ActorSystem, eventHelper: EventHelper, val schedulingProperties: SchedulingProperties)
-  extends EventPersister with Logging {
-  def persist(event: Event): Unit = {
-    eventHelper.newEvent(event)
-    log.info("Fake persisting event %s".format(event.externalId))
-  }
-  def persist(events: Seq[Event]): Unit = {
-    log.info("Fake persisting events %s".format(events map (_.externalId) mkString(",")))
-  }
 }
