+++ kifi-backend/modules/common/test/com/keepit/common/cache/TestCacheModule.scala
@@ -1,38 +1,21 @@
 package com.keepit.common.cache
 
 import com.google.inject.{Provides, Singleton}
-import scala.concurrent.duration._
 import com.keepit.common.social.{CommentWithBasicUserCache, BasicUserUserIdCache}
-import net.codingwell.scalaguice.ScalaModule
-import com.keepit.model.BookmarkUriUserCache
-import com.keepit.model.UserCollectionsCache
-import com.keepit.model.CollectionsForBookmarkCache
-import com.keepit.model.NormalizedURICache
-import com.keepit.model.NormalizedURIUrlHashCache
-import com.keepit.model.SocialUserInfoUserCache
-import com.keepit.model.SocialUserInfoNetworkCache
-import com.keepit.model.UnscrapableAllCache
-import com.keepit.model.UserExternalIdCache
-import com.keepit.model.UserIdCache
-import com.keepit.model.UserSessionExternalIdCache
-import com.keepit.model.ExternalUserIdCache
-import com.keepit.model.UserExperimentCache
-import com.keepit.model.SliderHistoryUserIdCache
-import com.keepit.model.BookmarkCountCache
-import com.keepit.model.SocialUserInfoCountCache
-import com.keepit.model.CommentCountUriIdCache
-import com.keepit.model.UserValueCache
-import com.keepit.model.BrowsingHistoryUserIdCache
-import com.keepit.model.ClickHistoryUserIdCache
+import com.keepit.model._
 import com.keepit.search.ActiveExperimentsCache
-import com.keepit.model.UserConnectionIdCache
+import scala.concurrent.duration._
 
-class DevCacheModule extends CacheModule(HashMapMemoryCacheModule()) {
-
+case class TestCacheModule() extends CacheModule(HashMapMemoryCacheModule()) {
   @Singleton
   @Provides
   def basicUserUserIdCache(outerRepo: FortyTwoCachePlugin) =
     new BasicUserUserIdCache((outerRepo, 7 days))
+
+  @Singleton
+  @Provides
+  def normalizedURIUrlHashCache(innerRepo: InMemoryCachePlugin, outerRepo: FortyTwoCachePlugin) =
+    new NormalizedURIUrlHashCache((innerRepo, 1 second), (outerRepo, 7 days))
 
   @Singleton
   @Provides
@@ -61,11 +44,6 @@ class DevCacheModule extends CacheModule(HashMapMemoryCacheModule()) {
 
   @Singleton
   @Provides
-  def normalizedURIUrlHashCache(outerRepo: FortyTwoCachePlugin) =
-    new NormalizedURIUrlHashCache((outerRepo, 7 days))
-
-  @Singleton
-  @Provides
   def socialUserInfoUserCache(outerRepo: FortyTwoCachePlugin) =
     new SocialUserInfoUserCache((outerRepo, 30 days))
 
@@ -77,7 +55,7 @@ class DevCacheModule extends CacheModule(HashMapMemoryCacheModule()) {
   @Singleton
   @Provides
   def unscrapableAllCache(outerRepo: FortyTwoCachePlugin) =
-    new UnscrapableAllCache((outerRepo, 0 second))
+    new UnscrapableAllCache((outerRepo, 30 days))
 
   @Singleton
   @Provides
@@ -112,12 +90,12 @@ class DevCacheModule extends CacheModule(HashMapMemoryCacheModule()) {
   @Singleton
   @Provides
   def bookmarkCountCache(outerRepo: FortyTwoCachePlugin) =
-    new BookmarkCountCache((outerRepo, 1 hour))
+    new BookmarkCountCache((outerRepo, 1 day))
 
   @Singleton
   @Provides
   def socialUserInfoCountCache(outerRepo: FortyTwoCachePlugin) =
-    new SocialUserInfoCountCache((outerRepo, 1 hour))
+    new SocialUserInfoCountCache((outerRepo, 1 day))
 
   @Singleton
   @Provides
@@ -149,3 +127,4 @@ class DevCacheModule extends CacheModule(HashMapMemoryCacheModule()) {
   def userConnectionIdCache(outerRepo: FortyTwoCachePlugin) =
     new UserConnectionIdCache((outerRepo, 7 days))
 }
+
