+++ kifi-backend/modules/search/app/com/keepit/search/SearchProdModule.scala
@@ -1,18 +1,20 @@
 package com.keepit.search
 
 import com.keepit.social.RemoteSecureSocialModule
-import com.keepit.common.cache.SearchCacheModule
+import com.keepit.common.cache.{EhCacheCacheModule, MemcachedCacheModule, SearchCacheModule}
 import com.keepit.common.{SearchBrowsingHistoryModule, SearchClickHistoryModule}
 import com.keepit.shoebox.ShoeboxServiceClientImplModule
+import com.keepit.module.DiscoveryImplModule
 
 case class SearchProdModule() extends SearchModule(
 
   // Common Functional Modules
-  cacheModule = SearchCacheModule(),
+  cacheModule = SearchCacheModule(MemcachedCacheModule(), EhCacheCacheModule()),
   secureSocialModule = RemoteSecureSocialModule(),
   shoeboxServiceClientModule = ShoeboxServiceClientImplModule(),
   clickHistoryModule = SearchClickHistoryModule(),
   browsingHistoryModule = SearchBrowsingHistoryModule(),
+  discoveryModule = DiscoveryImplModule(),
 
   // Search Functional Modules
   indexModule = IndexImplModule(),
