+++ kifi-backend/modules/search/test/com/keepit/search/MainSearcherTest.scala
@@ -126,7 +126,7 @@ class MainSearcherTest extends Specification {
 
   "MainSearcher" should {
     "search and categorize using social graph" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges)
@@ -179,7 +179,7 @@ class MainSearcherTest extends Specification {
     }
 
     "return a single list of hits" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges)
@@ -231,7 +231,7 @@ class MainSearcherTest extends Specification {
     }
 
     "search personal bookmark titles" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges, uniqueTitle = Some("personal title"))
@@ -289,7 +289,7 @@ class MainSearcherTest extends Specification {
     }
 
     "score using matches in a bookmark title and an article" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges, uniqueTitle = Some("personal title"))
@@ -318,7 +318,7 @@ class MainSearcherTest extends Specification {
     }
 
     "paginate" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges)
@@ -360,7 +360,7 @@ class MainSearcherTest extends Specification {
     }
 
     "boost recent bookmarks" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 1, numUris = 5)
         val Seq(user) = users
         val userId = user.id.get
@@ -389,7 +389,7 @@ class MainSearcherTest extends Specification {
     }
 
     "be able to cut the long tail" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 1, numUris = 10)
         val Seq(user) = users
         val userId = user.id.get
@@ -426,7 +426,7 @@ class MainSearcherTest extends Specification {
     }
 
     "show own private bookmarks" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 2, numUris = 20)
         val user1 = users(0)
         val user2 = users(1)
@@ -450,7 +450,7 @@ class MainSearcherTest extends Specification {
     }
 
     "not show friends private bookmarks" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 2, numUris = 20)
         val user1 = users(0)
         val user2 = users(1)
@@ -481,7 +481,7 @@ class MainSearcherTest extends Specification {
     }
 
     "search hits using a stemmed word" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 9, numUris = 9)
         val expectedUriToUserEdges = uris.toIterator.zip((1 to 9).iterator.map(users.take(_))).toList
         saveBookmarksByURI(expectedUriToUserEdges, uniqueTitle = Some("my books"))
@@ -507,7 +507,7 @@ class MainSearcherTest extends Specification {
     }
 
     "search within collections" in {
-      running(new DevApplication().withShoeboxServiceModule) {
+      running(new SearchApplication().withShoeboxServiceModule) {
         val (users, uris) = initData(numUsers = 2, numUris = 20)
         val user1 = users(0)
         val user2 = users(1)
