+++ kifi-backend/modules/shoebox/app/com/keepit/common/cache/ShoeboxCacheModule.scala
@@ -6,7 +6,7 @@ import scala.concurrent.duration._
 import com.keepit.search.ActiveExperimentsCache
 import com.keepit.common.social.{CommentWithBasicUserCache, BasicUserUserIdCache}
 
-case class ShoeboxCacheModule() extends CacheModule(MemcachedCacheModule(), EhCacheCacheModule()) {
+case class ShoeboxCacheModule(cachePluginModules: CachePluginModule*) extends CacheModule(cachePluginModules:_*) {
 
   @Singleton
   @Provides
