+++ kifi-backend/modules/shoebox/app/com/keepit/shoebox/ShoeboxProdModule.scala
@@ -1,6 +1,6 @@
 package com.keepit.shoebox
 
-import com.keepit.common.cache.ShoeboxCacheModule
+import com.keepit.common.cache.{EhCacheCacheModule, MemcachedCacheModule, ShoeboxCacheModule}
 import com.keepit.social.ShoeboxSecureSocialModule
 import com.keepit.search.SearchServiceClientImplModule
 import com.keepit.common.db.{DbInfo, SlickModule}
@@ -9,25 +9,28 @@ import com.keepit.common.social.SocialGraphImplModule
 import com.keepit.common.analytics.AnalyticsImplModule
 import com.keepit.learning.topicmodel.LdaTopicModelModule
 import com.keepit.model.SliderHistoryTrackerImplModule
-import com.keepit.common.mail.ShoeboxMailModule
+import com.keepit.common.mail.ProdMailModule
 import com.keepit.common.crypto.ShoeboxCryptoModule
 import com.keepit.common.store.ShoeboxS3Module
 import com.keepit.realtime.ShoeboxWebSocketModule
-import com.keepit.classify.DomainTagImporterImplModule
+import com.keepit.classify.ProdDomainTagImporterModule
 import scala.slick.session.{Database => SlickDatabase}
 import play.api.db.DB
 import play.api.Play
+import com.keepit.module.{DiscoveryImplModule, ProdActorSystemModule}
 
 case class ShoeboxProdModule() extends ShoeboxModule(
   // Common Functional Modules
-  cacheModule = ShoeboxCacheModule(),
+  cacheModule = ShoeboxCacheModule(MemcachedCacheModule(), EhCacheCacheModule()),
   secureSocialModule = ShoeboxSecureSocialModule(),
   searchServiceClientModule = SearchServiceClientImplModule(),
   clickHistoryModule = ShoeboxClickHistoryModule(),
   browsingHistoryModule = ShoeboxBrowsingHistoryModule(),
-  mailModule = ShoeboxMailModule(),
+  mailModule = ProdMailModule(),
   cryptoModule = ShoeboxCryptoModule(),
   s3Module = ShoeboxS3Module(),
+  actorSystemModule = ProdActorSystemModule(),
+  discoveryModule = DiscoveryImplModule(),
 
   // Shoebox Functional Modules
   slickModule = SlickModule(ShoeboxDbInfo()),
@@ -36,7 +39,7 @@ case class ShoeboxProdModule() extends ShoeboxModule(
   analyticsModule = AnalyticsImplModule(),
   webSocketModule = ShoeboxWebSocketModule(),
   topicModelModule = LdaTopicModelModule(),
-  domainTagImporterModule = DomainTagImporterImplModule(),
+  domainTagImporterModule = ProdDomainTagImporterModule(),
   sliderHistoryTrackerModule = SliderHistoryTrackerImplModule()
 )
